[SB/MRH][82391326] move event trigger to end of _generateErrors
For the current story, we're adding a feature-flipped Flight
component to AG that presents the zutron errors in the UI
differently. To accomplish that, we needed to move the trigger
in this script to the end of the method, so that the new
component can do its job after the normal error handling.