[SB][89443342] no more "SuccessError" events
The SuccessError stuff was really confusing and I couldn't find
anything (other than one place in AG that will be changed after this
is merged) that listened for any of those events.
Decided to remove "Success" from the event names.
Added some comments explaining why we do error handling in
the success callbacks.