[SB/HH][89611032] use GET so IE works
After spending a day and a half trying to understand this problem,
I'm going to give up and just use GET.

However I still don't understand why it's necessary to do so
since similar POST requests (such as the ones made by
_submitEmailRegistration and _submitLogin) work fine.

The problem I'm observing is that the parameters are lost or discarded
at some point server-side before the controller is called.
Again, it's very odd that the problem doesn't happen for
the SesssionsController (and possibly others).

As an experiment I even tried changing the url so that it POSTs
the password confirmation data to a simple Sinatra app (with similarly
configured rack-cors middleware), and the params were passed
into the route normally as one would expect. WTF!?